Rename TypeScript function utility to match canonical declaration
