+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/function/typeScriptFunction.ts
