+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/typed-datum/customTypedDatum.ts
@@ -1,4 +1,4 @@
-import { TypeScriptFunction } from '../function/function';
+import { TypeScriptFunction } from '../function/typeScriptFunction';
 import {
   TypeScriptArray,
   TypeScriptMap,
