+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/typed-datum/typeScriptTypedDatum.ts
@@ -1,4 +1,4 @@
-import { TypeScriptFunction } from '../function/function';
+import { TypeScriptFunction } from '../function/typeScriptFunction';
 import { TypeScriptObject } from '../object/object';
 
 export enum TypeScriptDatumTypeName {
